Remove URLSession-based favicons fetcher and FaviconsLinksExtractor
