+++ Sources/Bookmarks/FaviconsFetcher/BookmarkFaviconsMetadataStorage.swift
@@ -1,0 +1,79 @@
+//
+//  BookmarkFaviconsMetadataStorage.swift
+//  DuckDuckGo
+//
+//  Copyright © 2023 DuckDuckGo. All rights reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+import Foundation
+
+public protocol BookmarkFaviconsMetadataStoring {
+    func getBookmarkIDs() throws -> Set<String>
+    func storeBookmarkIDs(_ ids: Set<String>) throws
+
+    func getDomainsWithoutFavicon() throws -> Set<String>
+    func storeDomainsWithoutFavicon(_ domains: Set<String>) throws
+}
+
+public class BookmarkFaviconsMetadataStorage: BookmarkFaviconsMetadataStoring {
+
+    let dataDirectoryURL: URL
+    let missingIDsFileURL: URL
+    let domainsWithoutFaviconURL: URL
+
+    public init(applicationSupportURL: URL) {
+        dataDirectoryURL = applicationSupportURL.appendingPathComponent("FaviconsFetcher")
+        missingIDsFileURL = dataDirectoryURL.appendingPathComponent("missingIDs")
+        domainsWithoutFaviconURL = dataDirectoryURL.appendingPathComponent("domainsWithoutFavicon")
+
+        initStorage()
+    }
+
+    private func initStorage() {
+        if !FileManager.default.fileExists(atPath: dataDirectoryURL.path) {
+            try! FileManager.default.createDirectory(at: dataDirectoryURL, withIntermediateDirectories: true)
+        }
+        if !FileManager.default.fileExists(atPath: missingIDsFileURL.path) {
+            FileManager.default.createFile(atPath: missingIDsFileURL.path, contents: Data())
+        }
+        if !FileManager.default.fileExists(atPath: domainsWithoutFaviconURL.path) {
+            FileManager.default.createFile(atPath: domainsWithoutFaviconURL.path, contents: Data())
+        }
+    }
+
+    public func getBookmarkIDs() throws -> Set<String> {
+        let data = try Data(contentsOf: missingIDsFileURL)
+        guard let rawValue = String(data: data, encoding: .utf8) else {
+            return []
+        }
+        return Set(rawValue.components(separatedBy: ","))
+    }
+
+    public func storeBookmarkIDs(_ ids: Set<String>) throws {
+        try ids.joined(separator: ",").data(using: .utf8)?.write(to: missingIDsFileURL)
+    }
+
+    public func getDomainsWithoutFavicon() throws -> Set<String> {
+        let data = try Data(contentsOf: domainsWithoutFaviconURL)
+        guard let rawValue = String(data: data, encoding: .utf8) else {
+            return []
+        }
+        return Set(rawValue.components(separatedBy: ","))
+    }
+
+    public func storeDomainsWithoutFavicon(_ domains: Set<String>) throws {
+        try domains.joined(separator: ",").data(using: .utf8)?.write(to: domainsWithoutFaviconURL)
+    }
+}
